refactor(cli): extract shared runStagedFlows orchestrator
