+++ src/commands/flows/hybridRunDefaults.ts
@@ -1,53 +1,26 @@
 import { join, resolve } from "node:path";
 
 import { buildPatternArgs } from "~/core/patternArgs.js";
-import { runnerMessages } from "~/core/messages/index.js";
-import { resolveProjectDirSafe } from "~/domains/flows/ensureDeps.js";
 import { expandPatterns as defaultExpandPatterns } from "~/domains/flows/expand.js";
 import { handleFlowsPull } from "~/domains/flows/pull/handler.js";
 import { validateEnvId } from "~/domains/flows/pull/pull.js";
-import { createAndroidDeps } from "~/domains/runner/runAndroidFlowDeps.js";
 import { flowsRun as defaultFlowsRun } from "~/domains/runner/run.js";
 import type { FlowsRunFlags } from "~/domains/runner/runInternals.js";
 import { defaultRunWebFlowDeps } from "~/domains/runner/runWebFlowDeps.js";
-import type { EnsureRuntimeEnvResult } from "~/domains/runtimeEnv/index.js";
-import { managedEnvBaseDir } from "~/domains/runtimeEnv/managedEnvDir.js";
-import {
-  prepareRunDir as defaultPrepareRunDir,
-  type PrepareRunDirArgs,
-  type PrepareRunDirResult,
-} from "~/domains/runtimeEnv/prepareRunDir.js";
+import { prepareRunDir as defaultPrepareRunDir } from "~/domains/runtimeEnv/prepareRunDir.js";
 import type {
   AuthCommandContext,
   CommandResult,
 } from "~/shell/commandContext.js";
 import type { Fs } from "~/shell/fs.js";
-import type { Logger } from "~/shell/logger.js";
 import { configureTestkit as defaultConfigureTestkit } from "~/shell/testkit.js";
-import {
-  resolveDepsRoot,
-  type ResolveDepsRootArgs,
-} from "~/commands/resolveDepsRoot.js";
+import { resolveDepsRoot } from "~/commands/resolveDepsRoot.js";
 
-import { buildFlowsRunDeps } from "./buildFlowsRunDeps.js";
-import { loadEnvFile } from "./loadEnvFile.js";
+import { type HandleFlowsRunDeps } from "./runDefaults.js";
+import { runStagedFlows } from "./runStagedFlows.js";
 
-export type HandleHybridFlowsRunDeps = {
-  expandPatterns: (
-    patterns: string[],
-    cwd: string,
-    logger?: Logger,
-  ) => Promise<string[]>;
+export type HandleHybridFlowsRunDeps = HandleFlowsRunDeps & {
   pullEnv: (ctx: AuthCommandContext, envId: string) => Promise<CommandResult>;
-  resolveDepsRoot: (
-    args: Omit<ResolveDepsRootArgs, "fs">,
-  ) => Promise<EnsureRuntimeEnvResult>;
-  prepareRunDir: (
-    args: Omit<PrepareRunDirArgs, "fs">,
-  ) => Promise<PrepareRunDirResult>;
-  configureTestkit: (dir: string) => Promise<void>;
-  flowsRun: typeof defaultFlowsRun;
-  runWebFlowDeps: typeof defaultRunWebFlowDeps;
 };
 
 function makeDefaultHybridDeps(fs: Fs): HandleHybridFlowsRunDeps {
@@ -98,53 +71,5 @@ export async function handleHybridFlowsRun(
     }
   }
 
-  ctx.ui.gap();
-  ctx.ui.intro("flows run");
-
-  const [runtimeEnv] = await ctx.ui.withProgress(
-    [
-      {
-        message: runnerMessages.preparingEnvironment,
-        task: () =>
-          resolvedDeps.resolveDepsRoot({
-            files,
-            ...(flags.deps !== undefined ? { overrideDir: flags.deps } : {}),
-          }),
-      },
-    ],
-    () => runnerMessages.environmentReady,
-  );
-  if (runtimeEnv.source === "managed") {
-    ctx.ui.info(runnerMessages.managedRuntimeNote(runtimeEnv.depsRoot));
-  }
-  await loadEnvFile(envDir);
-
-  const projectDir = resolveProjectDirSafe(files, ctx.fs);
-  const staged = await resolvedDeps.prepareRunDir({
-    files,
-    projectDir,
-    depsRoot: runtimeEnv.depsRoot,
-    runRoot: join(managedEnvBaseDir(), ".runs"),
-  });
-
-  const resolvedDir = runtimeEnv.depsRoot;
-  await resolvedDeps.configureTestkit(resolvedDir);
-  const android = createAndroidDeps(resolvedDir, ctx.signals);
-  const runWebFlowDeps = await resolvedDeps.runWebFlowDeps(
-    resolvedDir,
-    ctx.signals,
-  );
-
-  const unregisterCleanup = ctx.signals.register(staged.cleanup);
-  try {
-    return await resolvedDeps.flowsRun(
-      ctx,
-      staged.files,
-      flags,
-      buildFlowsRunDeps({ ctx, resolvedDir, android, runWebFlowDeps, flags }),
-    );
-  } finally {
-    unregisterCleanup();
-    await staged.cleanup();
-  }
+  return runStagedFlows({ ctx, files, flags, envDir, deps: resolvedDeps });
 }
