+++ src/commands/flows/runStagedFlows.ts
@@ -1,0 +1,103 @@
+import { join } from "node:path";
+
+import { runnerMessages } from "~/core/messages/index.js";
+import { resolveProjectDirSafe } from "~/domains/flows/ensureDeps.js";
+import { createAndroidDeps } from "~/domains/runner/runAndroidFlowDeps.js";
+import type { flowsRun as defaultFlowsRun } from "~/domains/runner/run.js";
+import type { FlowsRunFlags } from "~/domains/runner/runInternals.js";
+import type { defaultRunWebFlowDeps } from "~/domains/runner/runWebFlowDeps.js";
+import {
+  type EnsureRuntimeEnvResult,
+  managedEnvBaseDir,
+} from "~/domains/runtimeEnv/index.js";
+import type {
+  PrepareRunDirArgs,
+  PrepareRunDirResult,
+} from "~/domains/runtimeEnv/prepareRunDir.js";
+import type { CommandContext, CommandResult } from "~/shell/commandContext.js";
+import type { ResolveDepsRootArgs } from "~/commands/resolveDepsRoot.js";
+
+import { buildFlowsRunDeps } from "./buildFlowsRunDeps.js";
+import { loadEnvFile } from "./loadEnvFile.js";
+
+/** Dependency bundle for the shared post-discovery run-setup phase. */
+export type StagedRunDeps = {
+  resolveDepsRoot: (
+    args: Omit<ResolveDepsRootArgs, "fs">,
+  ) => Promise<EnsureRuntimeEnvResult>;
+  prepareRunDir: (
+    args: Omit<PrepareRunDirArgs, "fs">,
+  ) => Promise<PrepareRunDirResult>;
+  configureTestkit: (dir: string) => Promise<void>;
+  runWebFlowDeps: typeof defaultRunWebFlowDeps;
+  flowsRun: typeof defaultFlowsRun;
+};
+
+export type RunStagedFlowsArgs = {
+  ctx: CommandContext;
+  files: string[];
+  flags: FlowsRunFlags;
+  // .env load dir override; defaults to the resolved project dir or cwd
+  envDir?: string;
+  deps: StagedRunDeps;
+};
+
+/**
+ * Shared run-setup for both local and hybrid flow runs. Handles environment
+ * resolution, staging, and execution after flow discovery is complete.
+ */
+export async function runStagedFlows(
+  args: RunStagedFlowsArgs,
+): Promise<CommandResult> {
+  const { ctx, files, flags, envDir, deps } = args;
+
+  ctx.ui.gap();
+  ctx.ui.intro("flows run");
+
+  const [runtimeEnv] = await ctx.ui.withProgress(
+    [
+      {
+        message: runnerMessages.preparingEnvironment,
+        task: () =>
+          deps.resolveDepsRoot({
+            files,
+            ...(flags.deps !== undefined ? { overrideDir: flags.deps } : {}),
+          }),
+      },
+    ],
+    () => runnerMessages.environmentReady,
+  );
+
+  if (runtimeEnv.source === "managed") {
+    ctx.ui.info(runnerMessages.managedRuntimeNote(runtimeEnv.depsRoot));
+  }
+
+  // Load the user's project .env from the project dir (NOT the deps dir).
+  const projectDir = resolveProjectDirSafe(files, ctx.fs);
+  await loadEnvFile(envDir ?? projectDir ?? process.cwd());
+
+  const staged = await deps.prepareRunDir({
+    files,
+    projectDir,
+    depsRoot: runtimeEnv.depsRoot,
+    runRoot: join(managedEnvBaseDir(), ".runs"),
+  });
+
+  const resolvedDir = runtimeEnv.depsRoot;
+  await deps.configureTestkit(resolvedDir);
+  const android = createAndroidDeps(resolvedDir, ctx.signals);
+  const runWebFlowDeps = await deps.runWebFlowDeps(resolvedDir, ctx.signals);
+
+  const unregisterCleanup = ctx.signals.register(staged.cleanup);
+  try {
+    return await deps.flowsRun(
+      ctx,
+      staged.files,
+      flags,
+      buildFlowsRunDeps({ ctx, resolvedDir, android, runWebFlowDeps, flags }),
+    );
+  } finally {
+    unregisterCleanup();
+    await staged.cleanup();
+  }
+}
